Таймер перед стартом уровня добавлен. Рефакторинг
