+++ TheStage/TheStage/Controls/GameContoller/PauseMenu.xaml
@@ -1,12 +1,11 @@
-﻿<UserControl x:Class="TheStage.Controls.PauseMenu"
+﻿<UserControl x:Class="TheStage.Controls.GameController.PauseMenu"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:ea="http://schemas.microsoft.com/expression/2010/interactions"
         xmlns:i="http://schemas.microsoft.com/expression/2010/interactivity"
-             xmlns:local="clr-namespace:TheStage.Controls"
-             xmlns:controls="clr-namespace:TheStage.Controls"
+             xmlns:controls="clr-namespace:TheStage.Controls.GameController"
              mc:Ignorable="d" 
              Background="#55000000"
              HorizontalAlignment="Stretch"
