+++ TheStage/TheStage/Controls/GameContoller/ShortStatistics.xaml
@@ -1,9 +1,8 @@
-﻿<UserControl x:Class="TheStage.Controls.ShortStatistics"
+﻿<UserControl x:Class="TheStage.Controls.GameController.ShortStatistics"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:TheStage.Controls"
              mc:Ignorable="d">
     <StackPanel>
         <StackPanel Orientation="Horizontal" Margin="5">
