+++ TheStage/TheStage/Controls/GameContoller/Statistics.xaml
@@ -1,10 +1,9 @@
-﻿<UserControl x:Class="TheStage.Controls.Statistics"
+﻿<UserControl x:Class="TheStage.Controls.GameController.Statistics"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:TheStage.Controls"
-             xmlns:control="clr-namespace:TheStage.Controls"
+             xmlns:control="clr-namespace:TheStage.Controls.GameController"
              mc:Ignorable="d"
              Background="#55000000"
              d:DesignHeight="600" d:DesignWidth="800">
@@ -23,7 +22,7 @@
 
             <control:ShortStatistics Grid.RowSpan="4"/>
             <TextBlock Grid.Column="1" Grid.Row="0" Text="{Binding Status}" Foreground="Red" HorizontalAlignment="Center" VerticalAlignment="Center" FontSize="20"/>
-            <Image Source="C:\Users\yuriy.medveditskov\Desktop\рамка_Сложности3.png" Grid.Row="1" Grid.Column="1" Width="75"/>
+            <Image Grid.Row="1" Grid.Column="1" Width="75"/>
             <TextBlock Margin="5" Grid.Column="1" Grid.Row="2" Text="{Binding Score}" Foreground="Red" HorizontalAlignment="Center" VerticalAlignment="Center" FontSize="20"/>
             <Grid Grid.Row="3" Grid.Column="1">
                 <Grid.ColumnDefinitions>
