+++ TheStage/TheStage/Controls/GameContoller/Timer.xaml
@@ -1,0 +1,34 @@
+﻿<UserControl x:Class="TheStage.Controls.GameContoller.Timer"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:TheStage.Controls.GameContoller"
+             mc:Ignorable="d"
+             Background="#AA000000"
+             HorizontalAlignment="Stretch"
+             VerticalAlignment="Stretch"
+             d:DesignHeight="720" d:DesignWidth="1280"
+             x:Name="ucTimer" x:FieldModifier="private">
+    <Grid Width="800" Height="600">
+        <TextBlock x:Name="txtTimer" x:FieldModifier="private" FontSize="400" Text="5" HorizontalAlignment="Center" VerticalAlignment="Center" TextAlignment="Center" Foreground="AliceBlue">
+            <TextBlock.Triggers>
+                <EventTrigger RoutedEvent="TextBlock.Loaded">
+                    <BeginStoryboard>
+                        <Storyboard>
+                            <StringAnimationUsingKeyFrames Storyboard.TargetName="txtTimer" Storyboard.TargetProperty="Text" >
+                                <DiscreteStringKeyFrame Value="5" KeyTime="0:0:0"/>
+                                <DiscreteStringKeyFrame Value="4" KeyTime="0:0:1"/>
+                                <DiscreteStringKeyFrame Value="3" KeyTime="0:0:2"/>
+                                <DiscreteStringKeyFrame Value="2" KeyTime="0:0:3"/>
+                                <DiscreteStringKeyFrame Value="1" KeyTime="0:0:4"/>
+                                <DiscreteStringKeyFrame Value="0" KeyTime="0:0:5"/>
+                            </StringAnimationUsingKeyFrames>
+                            <DoubleAnimation Storyboard.TargetName="ucTimer" Storyboard.TargetProperty="Opacity" From="1" To="0" BeginTime="0:0:4" Duration="0:0:1" Completed="DoubleAnimation_Completed"/>
+                        </Storyboard>
+                    </BeginStoryboard>
+                </EventTrigger>
+            </TextBlock.Triggers>
+        </TextBlock>
+    </Grid>
+</UserControl>
